sink: Move the homegrown congestion controller in it own file
